requirements.txt
requests
pandas
beautifulsoup4
pyyaml
